+++ validator.js
@@ -1393,7 +1393,7 @@
         return parts.join('@');
       }
 
-      var version = '6.0.0';
+      var version = '6.1.0';
 
       var validator = {
         version: version,