+++ CHANGELOG.md
@@ -4,7 +4,8 @@
   ([#585](https://github.com/chriso/validator.js/pull/585))
 - Added support for greater or less than in `isFloat()`
   ([#544](https://github.com/chriso/validator.js/issues/544))
-- Added support for `lt` and `gt` into `isInt()`.
+- Added support for `lt` and `gt` into `isInt()`
+  ([#588](https://github.com/chriso/validator.js/pull/588))
 
 #### 6.0.0